Getting the food sprite in
Setting up the food sprite. I dont have all the data for it yet, that will take some doing. Basics are in though. Also adjusted the formatting of each sprite to center the title and pretty it up a bit. Also changed Reactors to Nuclear Reactors from a player facing perspective