Update LCDInfo.cs
DA BIG UPDATE. This has some of the formatting cleaned up around the new sprites, and also a Grid Info sprite added in. I want to expand on the grid info sprite in future with things like grid mass, etc...

This is also the version of the mod that is going live!